requirements.txt
streamlit
numpy
trimesh
openai
python-dotenv
fastapi
uvicorn
python-multipart
scipy
